Add relay rack ips as DNS servers
This patch aims to add the rack controll IP addresses of the relay
vlan as DNS servers.
This fixes an issue that when there are no other settings or available
IPs, MAAS adds only the IP address of the region controller, but
sometimes the region controller might not be routed to that vlan, hence
this patch, as the rack controllers should always be reachable by the
machines.

Signed-off-by: David Negreira <david.negreira@canonical.com>